Engine lift method fix

+++ src/AutomatedCar/Models/PowerTrain/DasboardModel.cs
@@ -32,7 +32,7 @@ namespace AutomatedCar.Models.PowerTrain
         {
             get
             {
-                return this.rpm;
+                return (int)this.rpm;
             }
 
             set
